Transform patterns from relative form even if are not valid in FS
The logic to process patterns to determine include paths is separated
from the logic to process patterns that remain as globs.

patterns that remain as globs are transformed from relative form to
absolute form and keeping the suffix, these don't need to be validated
against the file sytem.

include patters that become include paths need to be validated against
the file system.
Additionally when resolving include paths, don't transform include
patterns not given in relative form to absolute form, so the search
can be broad and only become specfic when in relative form.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>