Don't re-open hidden tree-view after reload
The PluginViewWidget has a property to suppress visibility updates
from a `when` clause in the case a user specifically requests to hide
it. However when the browser reloaded, the flag was reset to default
and the prvided `when` clause would apply to decide the visibility.

This change preserves the flag mentioned above in the PluginViewWidget
state, this state is presered and then available for use in
`restoreState` where the flag is now updated to the state before the
browser reload.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>